Add header to swagger and revert changes in autogenerated file
