+++ pkg/services/ngalert/api/provisioning.go
@@ -4,7 +4,6 @@ import (
 	"github.com/grafana/grafana/pkg/api/response"
 	"github.com/grafana/grafana/pkg/models"
 	apimodels "github.com/grafana/grafana/pkg/services/ngalert/api/tooling/definitions"
-	alerting_models "github.com/grafana/grafana/pkg/services/ngalert/models"
 )
 
 type ProvisioningApiHandler struct {
@@ -81,12 +80,12 @@ func (f *ProvisioningApiHandler) handleRouteGetAlertRule(ctx *models.ReqContext,
 	return f.svc.RouteRouteGetAlertRule(ctx, UID)
 }
 
-func (f *ProvisioningApiHandler) handleRoutePostAlertRule(ctx *models.ReqContext, ar apimodels.ProvisionedAlertRule, provenance alerting_models.Provenance) response.Response {
-	return f.svc.RoutePostAlertRule(ctx, ar, provenance)
+func (f *ProvisioningApiHandler) handleRoutePostAlertRule(ctx *models.ReqContext, ar apimodels.ProvisionedAlertRule) response.Response {
+	return f.svc.RoutePostAlertRule(ctx, ar)
 }
 
-func (f *ProvisioningApiHandler) handleRoutePutAlertRule(ctx *models.ReqContext, ar apimodels.ProvisionedAlertRule, UID string, provenance alerting_models.Provenance) response.Response {
-	return f.svc.RoutePutAlertRule(ctx, ar, UID, provenance)
+func (f *ProvisioningApiHandler) handleRoutePutAlertRule(ctx *models.ReqContext, ar apimodels.ProvisionedAlertRule, UID string) response.Response {
+	return f.svc.RoutePutAlertRule(ctx, ar, UID)
 }
 
 func (f *ProvisioningApiHandler) handleRouteDeleteAlertRule(ctx *models.ReqContext, UID string) response.Response {
